Update markup + improve handling project changes
Updated the markup to use a collection of paragraphs. This is so that it
can take advantage of the built in browser CSS columns. This then lets
us hide or show projects in the list and have the columns dynamically
reflow.

Updated the JavaScript to no longer store Jenkins views as it is
simplier to work out which projects to remove from view.

Also ensured that projects always maintain alphabetic ordering.